Rewrite container logic with CSS grid
