+++ src/css/main.css
@@ -51,7 +51,7 @@ body{margin:0;font-family:sans-serif}
   position:absolute;right:2px;bottom:2px;width:12px;height:12px;
   background:#007bff50;border-radius:2px;cursor:se-resize;
 }
-.subgrid .resize-handle{display:none}
+.native-grid .resize-handle{display:none}
 
 .container{transition:min-height .3s ease;width:100%;}
 .container .collapse__header{display:flex;align-items:center;gap:.25rem}
@@ -60,8 +60,9 @@ body{margin:0;font-family:sans-serif}
 .container .collapse__body{overflow-y:auto;overflow-x:hidden;padding:8px}
 .container .collapse__body::-webkit-scrollbar{width:8px}
 .container .collapse__body::-webkit-scrollbar-thumb{background:#007bff;border-radius:4px}
-.container .subgrid{min-height:100px;width:100%}
-.container .subgrid>.grid-stack-item{min-width:200px;max-width:400px}
+.container .card-wrapper{width:100%;height:100%;max-height:700px;display:flex;flex-wrap:wrap}
+.container .native-grid{display:grid;gap:16px;grid-template-columns:repeat(auto-fit,minmax(300px,1fr));grid-auto-rows:200px;min-height:100px;max-height:700px}
+.container .native-grid>[gs-id]{min-width:300px;max-width:400px;min-height:200px;max-height:200px}
 .container.collapsed{min-height:100px;height:100px;overflow:hidden;position:relative}
 .container.collapsed .collapse__body{display:none}
 .container.collapsed::after{
